+++ src/game/tests.rs
@@ -588,6 +588,7 @@ mod old_school_permanents;
 mod old_school_spells;
 mod old_school_upkeep;
 mod one_sided_block_triggers;
+mod open_activation;
 mod poison;
 mod populate;
 mod power_blocking;